requirements.txt
# SlipScan — Python Dependencies (ทั้งระบบ)
# ติดตั้งครั้งเดียว: pip install -r requirements.txt

# ── Flask (OCR Service + Backend) ─────────────────────────────────────────────
Flask==3.1.0
flask-cors==5.0.1
Werkzeug==3.1.6
python-dotenv==1.0.1
itsdangerous==2.2.0
click==8.3.1
blinker==1.9.0
Jinja2==3.1.6
MarkupSafe==3.0.3
colorama==0.4.6

# ── Backend Auth ───────────────────────────────────────────────────────────────
PyJWT==2.10.1
bcrypt==4.3.0

# ── Database (Supabase PostgreSQL) ────────────────────────────────────────────
psycopg2-binary==2.9.10

# ── Typhoon OCR Engine ─────────────────────────────────────────────────────────
typhoon-ocr==0.4.1
openai==2.21.0
httpx==0.28.1
httpcore==1.0.9
pydantic==2.12.5
pydantic_core==2.41.5
annotated-types==0.7.0
anyio==4.12.1
sniffio==1.3.1
jiter==0.13.0
distro==1.9.0
typing_extensions==4.15.0
typing-inspection==0.4.2

# ── Image Preprocessing ────────────────────────────────────────────────────────
Pillow==12.1.1
opencv-python-headless==4.13.0.92
numpy==2.2.6
ftfy==6.3.1
wcwidth==0.6.0

# ── PDF Support ────────────────────────────────────────────────────────────────
pypdf==6.7.2

# ── HTTP / Utilities ───────────────────────────────────────────────────────────
requests==2.32.3
certifi==2026.2.25
charset-normalizer==3.4.4
idna==3.11
urllib3==2.6.3
tqdm==4.67.3
h11==0.16.0
